Centralizing Uuid generation, validation
